requirements.txt
django==4.2.8
python-decouple
djangorestframework
djangorestframework-simplejwt==5.3.1
django-redis
python-dotenv==1.0.1
gunicorn
psycopg2-binary
Pillow
django-modeltranslation
drf-yasg
django-cors-headers
django-filter
python-dateutil
ujson
channels_redis
requests
sentry-sdk[django]
django-debug-toolbar
drf-standardized-errors==0.13.0